\name{neat-package}
\alias{neat-package}
\docType{package}
\title{neat}
\description{Includes functions and examples to compute NEAT, the Network Enrichment Analysis Test described in Signorelli et al. (2016).}
\author{Mirko Signorelli}
\references{
Signorelli, M., Vinciotti, V., Wit, E. C. (2016). NEAT: an efficient network enrichment analysis test. BMC Bioinformatics, 17:352. Url: https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-016-1203-6.
}
\keyword{package}
\seealso{\code{\link{neat}}}
