\name{neat-package}
\alias{neat-package}
\docType{package}
\title{neat}
\description{Includes functions and examples to compute NEAT (Network Enrichment Analysis Test), a network-based test for genetic enrichment analysis (Signorelli et al., 2016).}
\author{Mirko Signorelli}
\references{
Signorelli, M., Vinciotti, V., Wit, E. C. (2016). NEAT: an efficient network enrichment analysis test. ArXiv preprint, arXiv:1604.01210. Url: https://arxiv.org/abs/1604.01210.
}
\keyword{package}
\seealso{\code{\link{neat}}}
